Split up the window functions/helpers/events some
